+++ Classes/Domain/Model/Country.php
@@ -7,7 +7,7 @@ namespace SJBR\StaticInfoTables\Domain\Model;
  *  (c) 2011-2012 Armin Rüdiger Vieweg <info@professorweb.de>
  *  (c) 2013 Stanislas Rolland <typo3(arobas)sjbr.ca>
  *  (c) 2015 Tomasz Krawczyk <tomasz(na)typo3.pl>
- *  (c) 2021-2025 Ephraim Härer <ephraim.haerer@renolit.com>
+ *  (c) 2021-2026 Ephraim Härer <ephraim.haerer@renolit.com>
  *  All rights reserved
  *  This script is part of the TYPO3 project. The TYPO3 project is
  *  free software; you can redistribute it and/or modify