+++ Configuration/TCA/Overrides/static_currencies.php
@@ -1,19 +1,17 @@
 <?php
-defined('TYPO3') || die();
+declare(strict_types=1);
 
-call_user_func(function ($extKey, $table) {
-    $additionalFields = [
+defined('TYPO3') || die;
+
+use RENOLIT\StaticInfoTablesPl\Provider\TcaProvider;
+
+call_user_func(
+    function ($additionalFields, $dataSetName) {
+        TcaProvider::generateAndRegisterTca($additionalFields, $dataSetName);
+    },
+    [
         'cu_name_en' => 'cu_name_pl',
         'cu_sub_name_en' => 'cu_sub_name_pl'
-    ];
-    foreach ($additionalFields as $sourceField => $destField) {
-        $additionalColumns = [];
-        $additionalColumns[$destField] = $GLOBALS['TCA'][$table]['columns'][$sourceField];
-        $additionalColumns[$destField]['label'] = 'LLL:EXT:' . $extKey . '/Resources/Private/Language/locallang_db.xlf:static_currencies_item.' . $destField;
-        \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addTCAcolumns($table, $additionalColumns);
-        \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addToAllTCAtypes($table, $destField, '',
-            'after:' . $sourceField);
-        // Add as search field
-        $GLOBALS['TCA'][$table]['ctrl']['searchFields'] .= ',' . $destField;
-    }
-}, 'static_info_tables_pl', 'static_currencies');
+    ],
+    'static_info_tables_pl'
+);