+++ ext_emconf.php
@@ -11,12 +11,13 @@
 $EM_CONF[$_EXTKEY] = [
     'title' => 'Static Info Tables (pl)',
     'description' => 'Polish (pl) language pack for the Static Info Tables providing localized names for countries, currencies and so on.',
-    'version' => '13.4.0',
+    'version' => '14.3.0',
     'category' => 'misc',
     'constraints' => [
         'depends' => [
-            'typo3' => '13.4.0-13.4.99',
-            'static_info_tables' => '13.4.0-13.4.99',
+            'typo3' => '14.3.0-14.3.99',
+            'php' => '8.2.0-8.5.99',
+            'static_info_tables' => '14.3.0-14.3.99',
         ],
         'conflicts' => [],
         'suggests' => [],
@@ -25,5 +26,10 @@ $EM_CONF[$_EXTKEY] = [
     'author' => 'Ephraim Härer',
     'author_email' => 'ephraim.haerer@renolit.com',
     'author_company' => 'RENOLIT SE',
+    'autoload' => [
+        'psr-4' => [
+            'RENOLIT\\StaticInfoTablesPl\\' => 'Classes',
+        ],
+    ],
 ];