+++ ChangeLog.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+### 2026-07-23  Ephraim Härer  <ephraim.haerer@renolit.com>
+- new version for TYPO3 14.3
+- updated classes, cleanup of PHP files
+- updated for TYPO3 14.3
+
+
 ### 2025-01-15  Ephraim Härer  <ephraim.haerer@renolit.com>
 - updated for TYPO3 13.4